tests/wpt/web-platform-tests/css/work-in-progress/gadams/cssom/csscharsetrule-interface.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD//XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml">
  <head>
    <title>CSSOM - CSSCharsetRule - Interface</title>
    <link rel="author" title="Glenn Adams" href="mailto:glenn@skynav.com"/>
    <link rel="help" href="http://www.w3.org/TR/cssom-1/#the-csscharsetrule-interface"/>
    <meta name="flags" content="dom"/>
    <script src="/resources/testharness.js"/>
    <script src="/resources/testharnessreport.js"/>
  </head>
  <body>
    <div id="log"></div>
    <div id="box"></div>
    <script><![CDATA[

      test(function(){

        assert_true('CSSCharsetRule' in window);

      }, '\'CSSCharsetRule\' in window');

      test(function(){

        assert_true(CSSCharsetRule instanceof Function);

      }, 'CSSCharsetRule instanceof Function');

      test(function(){

        assert_true('prototype' in CSSCharsetRule);

      }, '\'prototype\' in CSSCharsetRule');

      test(function(){

        assert_equals(CSSCharsetRule.prototype.toString(), '[object CSSCharsetRulePrototype]');

      }, 'CSSCharsetRule.prototype.toString() == \'[object CSSCharsetRulePrototype]\'');

      test(function(){

        assert_true('constructor' in CSSCharsetRule.prototype);

      }, '\'constructor\' in CSSCharsetRule.prototype');

      test(function(){

        assert_equals(CSSCharsetRule.prototype.constructor,CSSCharsetRule);

      }, 'CSSCharsetRule.prototype.constructor === CSSCharsetRule');

    ]]>
    </script>
  </body>
</html>
